+++ SisprodIT2/Areas/Perfil/Controllers/PerfilController.cs
@@ -11,6 +11,7 @@ using SisprodIT2.Map;
 namespace SisprodIT2.Areas.Perfil.Controllers
 {
     [Authorize]
+    [SessionExpire]
     public class PerfilController : Controller
     {
         private DataContext db = new DataContext();
@@ -20,6 +21,10 @@ namespace SisprodIT2.Areas.Perfil.Controllers
 
         public ActionResult Index()
         {
+            if (!Session["Perfil"].ToString().Equals("Administrador"))
+            {
+                return RedirectToAction("SemPermissao", "Home", new { area = "" });
+            }
             return View(db.Perfis.ToList());
         }
 
@@ -28,6 +33,10 @@ namespace SisprodIT2.Areas.Perfil.Controllers
 
         public ActionResult Details(int id = 0)
         {
+            if (!Session["Perfil"].ToString().Equals("Administrador"))
+            {
+                return RedirectToAction("SemPermissao", "Home", new { area = "" });
+            }
             PerfilModel perfilmodel = db.Perfis.Find(id);
             if (perfilmodel == null)
             {
@@ -41,6 +50,10 @@ namespace SisprodIT2.Areas.Perfil.Controllers
 
         public ActionResult Create()
         {
+            if (!Session["Perfil"].ToString().Equals("Administrador"))
+            {
+                return RedirectToAction("SemPermissao", "Home", new { area = "" });
+            }
             ViewBag.FuncionarioModelId = Session["FuncionarioModelId"];
             return View();
         }
@@ -51,6 +64,10 @@ namespace SisprodIT2.Areas.Perfil.Controllers
         [HttpPost]
         public ActionResult Create(PerfilModel perfilmodel)
         {
+            if (!Session["Perfil"].ToString().Equals("Administrador"))
+            {
+                return RedirectToAction("SemPermissao", "Home", new { area = "" });
+            }
             if (ModelState.IsValid)
             {
                 db.Perfis.Add(perfilmodel);
@@ -66,6 +83,11 @@ namespace SisprodIT2.Areas.Perfil.Controllers
 
         public ActionResult Edit(int id = 0)
         {
+            if (!Session["Perfil"].ToString().Equals("Administrador"))
+            {
+                return RedirectToAction("SemPermissao", "Home", new { area = "" });
+            }
+            ViewBag.FuncionarioModelId = Session["FuncionarioModelId"];
             PerfilModel perfilmodel = db.Perfis.Find(id);
             if (perfilmodel == null)
             {
@@ -80,6 +102,10 @@ namespace SisprodIT2.Areas.Perfil.Controllers
         [HttpPost]
         public ActionResult Edit(PerfilModel perfilmodel)
         {
+            if (!Session["Perfil"].ToString().Equals("Administrador"))
+            {
+                return RedirectToAction("SemPermissao", "Home", new { area = "" });
+            }
             if (ModelState.IsValid)
             {
                 db.Entry(perfilmodel).State = EntityState.Modified;
@@ -94,7 +120,12 @@ namespace SisprodIT2.Areas.Perfil.Controllers
 
         public ActionResult Delete(int id = 0)
         {
+            if (!Session["Perfil"].ToString().Equals("Administrador"))
+            {
+                return RedirectToAction("SemPermissao", "Home", new { area = "" });
+            }
             PerfilModel perfilmodel = db.Perfis.Find(id);
+
             if (perfilmodel == null)
             {
                 return HttpNotFound();
@@ -108,6 +139,10 @@ namespace SisprodIT2.Areas.Perfil.Controllers
         [HttpPost, ActionName("Delete")]
         public ActionResult DeleteConfirmed(int id)
         {
+            if (!Session["Perfil"].ToString().Equals("Administrador"))
+            {
+                return RedirectToAction("SemPermissao", "Home", new { area = "" });
+            }
             PerfilModel perfilmodel = db.Perfis.Find(id);
             db.Perfis.Remove(perfilmodel);
             db.SaveChanges();
